+++ CHANGELOG.md
@@ -5,6 +5,8 @@ principles.
 ## next release
 - Add: option to set custom from email address. Via chief setting from_email and from_name values
 - Added: sidebar component::onComponentReloading and component::onComponentReloaded listeners.
+- Added: option to disable livewire reload after sidebar component submission by passing livewire=false in response data.
+- Added: response data to sidebar event payload data
 - Removed: sidebar component::onComponentReload event listener. use the onComponentReloaded instead
 
 ## 0.6.4 - 2021-09-24